added code to handle exceptions during parsing in the selectstatement constructor added code to properly handle columnname tokens by correctly setting data_row_key and row variables Fixed inefficient writeline code where QTextStream was being created on every writeline

+++ Types.h
@@ -95,7 +95,7 @@ namespace csvquery {
 
 
 
-    extern QHash<QString, double> columns_table; // {COLUMNNANME, index}
+    extern QHash<QString, int> columns_table; // {COLUMNNANME, index}
 
     extern QHash<QString, double> numbers_table; // {name, number}
